Crate icu_unicodeset

UnicodeSet

This crate is the ICU4X implementation of the existing ICU4C UnicodeSet API.

Purpose

UnicodeSet is designed to allow for highly efficient querying of sets of Unicode characters.

Architecture

ICU4X UnicodeSet is split up into independent levels, with UnicodeSet representing the membership/query API, and UnicodeSetBuilder representing the builder API. A Properties API is in future works.

Examples:

Creating a UnicodeSet

UnicodeSets are created from either serialized UnicodeSets, represented by inversion lists, the UnicodeSetBuilder, or from the TBA Properties API.

use icu_unicodeset::UnicodeSet;
let inv_list = vec![65, 70]; // 'A'..'F'
let set = UnicodeSet::from_inversion_list(inv_list);
assert!(set.is_ok());

use icu_unicodeset::UnicodeSetBuilder;
let mut builder = UnicodeSetBuilder::new();
builder.add_range(&('A'..'Z'));
let set: UnicodeSet = builder.build();
assert!(set.contains('A'));

Querying the UnicodeSet

Currently, you can check if a character/range of characters exists in the UnicodeSet, or iterate through the characters.

use icu_unicodeset::UnicodeSet;
let inv_list = vec![65, 70];  // 'A'..'F'
let set = UnicodeSet::from_inversion_list(inv_list).unwrap();
assert!(set.contains('A'));
assert!(set.contains_range(&('A'..='C')));
assert_eq!(set.iter().next(), Some('A'));

Structs

UnicodeSet
UnicodeSet membership wrapper

UnicodeSetBuilder
UnicodeSet builder wrapper

Enums

UnicodeSetError
Custom Errors for UnicodeSet.

UnicodeSetSpanCondition

Functions

deconstruct_range
Returns start (inclusive) and end (excluisive) bounds of RangeBounds

is_valid
Returns whether the vector is sorted ascending non inclusive, of even length, and within the bounds of 0x0 -> 0x10FFFF inclusive.